moved event components to event folder and fixed the tab bar on view events page
