+++ lib/view/components/event/joined_event_card.dart
